+++ src/Parametizer/Script/ScriptLauncher/ScriptLauncher.php
@@ -63,7 +63,7 @@ class ScriptLauncher {
             : null;
 
         $classNamesBySubcommandNames = $this->scriptDetector
-            ->getFQClassNamesByScriptNames();
+            ->getClassNamesByScriptNames();
 
         // Init a subcommand switch - ensure built-in subcommands are added even if no custom subcommands are detected.
         $this->configBuilder->newSubcommandSwitch('subcommand-name');